fix: check if the user story has already been created in the issue tracker, and if not, set a custom response and generate a new info message

+++ frontend/src/views/SessionPage.vue
@@ -396,7 +396,11 @@ export default Vue.extend({
       if (this.session_userStoryMode === "US_JIRA") {
         let response;
         if (doRemove) {
-          response = await apiService.deleteUserStory(us[idx].id);
+          if (us[idx].id === null) {
+            response = 204;
+          } else {
+            response = await apiService.deleteUserStory(us[idx].id);
+          }
           us.splice(idx, 1);
           doRemove = false;
         } else {
@@ -420,6 +424,8 @@ export default Vue.extend({
         }
         if (response.status === 200) {
           this.$toast.success(this.$t("session.notification.messages.issueTrackerSynchronizeSuccess"));
+        } else if (response === 204) {
+          this.$toast.info(this.$t("session.notification.messages.issueTrackerSynchronizeNothingChanged"));
         } else {
           this.$toast.error(this.$t("session.notification.messages.issueTrackerSynchronizeFailed"));
         }
